Instead of just trusting randomness of authentication cookies between auth<->master<->login process IPC, master now doesn't accept any cookies from login process which weren't created by it (identified by PID). When login process dies, all it's pending cookies are also removed, so I can't see even a theoretical possiblity anymore for exploited login process to authenticate as someone else.
Also fixed some int -> unsigned int.

--HG--
branch : HEAD

+++ src/master/common.h
@@ -44,6 +44,6 @@ void imap_process_destroyed(pid_t pid);
 /* misc */
 #define VALIDATE_STR(str) \
 	validate_str(str, sizeof(str))
-int validate_str(const char *str, int max_len);
+int validate_str(const char *str, size_t max_len);
 
 #endif
